contributors/gadams/incoming/cssom/cssstyledeclaration-csstext-000.htm

<!DOCTYPE html>
<html>
  <head>
    <title>CSSOM - CSSStyleDeclaration.cssText - Delimiter Serialization.</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com">
    <link rel="help" href="http://www.w3.org/TR/cssom/#the-cssstyledeclaration-interface">
    <meta name="flags" content="dom">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script>
      var style = document.getElementById('box').style;

      test(function(){

        style.left = "10px";
        assert_equals(style.cssText, "left: 10px; ");

      }, 'CSSStyleDeclaration.cssText - delimiters - one declaration.');

      test(function(){

        style.left = "10px";
        style.right = "20px";
        assert_equals(style.cssText, "left: 10px; right: 20px; ");

      }, 'CSSStyleDeclaration.cssText - delimiters - two declarations.');
    </script>
  </body>
</html>
